Refactor: split ui into several files
